requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
onnx>=1.14.0
onnxruntime>=1.15.0
tokenizers>=0.13.0
huggingface_hub
python-dotenv
scikit-learn
